requirements.txt
pydantic
requests
instructor
openai
python-dotenv
pandas
google-cloud-bigquery

